# api/openapi.yaml
openapi: 3.0.3
info:
  title: Inroad API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /auth/api-keys:
    post:
      operationId: authApiKeyCreate
      tags: [auth]
      security: [{ bearerAuth: [] }]
      description: >-
        Mint a scoped API key for the caller's workspace. The full token
        (inrd_<prefix>_<secret>) is returned in this response EXACTLY ONCE and is
        never retrievable again — only its SHA-256 is stored. Grant a subset of the
        owned scope vocabulary; optionally restrict by CIDR allowlist, a per-minute
        rate limit, and an expiry. Session-authed only (an API key cannot mint keys).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApiKeyCreateRequest' }
      responses:
        '201':
          description: The created key metadata plus the one-time token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKeyCreateResponse' }
        '400': { description: Validation error (unknown scope, bad CIDR, past expiry, non-positive rate limit) }
        '401': { description: Unauthorized }
    get:
      operationId: authApiKeyList
      tags: [auth]
      security: [{ bearerAuth: [] }]
      description: >-
        List the caller's workspace API keys, newest first. No secret or hash is
        ever returned — only the public prefix and metadata.
      responses:
        '200':
          description: The workspace's API keys
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKeyListResponse' }
        '401': { description: Unauthorized }
  /auth/api-keys/{id}:
    delete:
      operationId: authApiKeyRevoke
      tags: [auth]
      security: [{ bearerAuth: [] }]
      description: >-
        Revoke one of the caller's workspace API keys. Idempotent (re-revoking
        succeeds); a foreign or unknown id returns 404 so a caller cannot probe or
        revoke another workspace's key.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '204': { description: Key revoked }
        '400': { description: Invalid id }
        '401': { description: Unauthorized }
        '404': { description: API key not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ApiKeyCreateRequest:
      type: object
      required: [name, scopes]
      properties:
        name: { type: string, description: Operator-facing label for the key. }
        scopes:
          type: array
          minItems: 1
          description: >-
            Granted scopes; a subset of the owned vocabulary (e.g. mailboxes:read,
            campaigns:send). The key holds exactly these.
          items: { type: string }
        ip_allowlist:
          type: array
          nullable: true
          description: >-
            Optional IPs or CIDRs the key may be used from; a bare IP is treated as
            a host route. Omit or leave empty for no restriction.
          items: { type: string }
        rate_limit_per_min:
          type: integer
          nullable: true
          description: Optional per-minute request cap. Omit for unlimited.
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: Optional RFC3339 expiry. Omit for a key that never expires.
    ApiKey:
      type: object
      description: Non-secret view of an API key. Never carries the secret or its hash.
      required: [id, name, prefix, scopes, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        prefix: { type: string, description: Public, non-secret token prefix (the part before the secret). }
        scopes: { type: array, items: { type: string } }
        ip_allowlist: { type: array, nullable: true, items: { type: string } }
        rate_limit_per_min: { type: integer, nullable: true }
        expires_at: { type: string, format: date-time, nullable: true }
        revoked_at: { type: string, format: date-time, nullable: true }
        last_used_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    ApiKeyCreateResponse:
      type: object
      required: [token, api_key]
      properties:
        token:
          type: string
          description: >-
            The full API key token (inrd_<prefix>_<secret>). SHOWN EXACTLY ONCE —
            store it now; it is never retrievable again.
        api_key: { $ref: '#/components/schemas/ApiKey' }
    ApiKeyListResponse:
      type: object
      required: [api_keys]
      properties:
        api_keys:
          type: array
          items: { $ref: '#/components/schemas/ApiKey' }
...